+++ src/scippneutron/conversion/beamline.py
