+++ src/scippneutron/conversion/tof.py
